// src/app/modulos/ventas/models/producto-response.model.ts
export class ProductoResponse {
    idProducto: number=0;
    idCategoria: number=0;
    idImagen: number=0;
    idTipoProducto:number=0;  
    nombre: string="";
    descripcion: string="";
    precio: number=0;
    stock: number=0;
    esUniversal: boolean=false;
    codigoProducto: string="";
    imagenUrl:string=""
}